# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthNonce {
  nonce: string;
}

export interface SiweVerification {
  message: string;
  signature: string;
}

export interface Player {
  walletAddress: string;
  xp: number;
  gold: number;
  level: number;
  selectedGoal?: string | null;
}

export interface PlayerGoalUpdate {
  selectedGoal: string | null;
}

export interface AuthUser {
  walletAddress: string;
  player: Player;
}

export interface AuthSession {
  user: AuthUser | null;
}

export interface PlayerAdjustment {
  goldDelta?: number;
  xpDelta?: number;
}

export interface PlayerItem {
  instanceId: string;
  itemId: string;
  slot: string;
  equipped: boolean;
  obtainedAt: string;
}

export interface PlayerItemInput {
  itemId: string;
  slot: string;
}

export interface PlayerItemUpdate {
  equipped: boolean;
}

export type ArenaMatchInputResult = typeof ArenaMatchInputResult[keyof typeof ArenaMatchInputResult];


export const ArenaMatchInputResult = {
  win: 'win',
  loss: 'loss',
  draw: 'draw',
} as const;

export interface ArenaMatchInput {
  opponentId: string;
  opponentName: string;
  result: ArenaMatchInputResult;
  xpEarned?: number;
  goldEarned?: number;
}

export interface ArenaMatchResult {
  ok: boolean;
}

export interface RankedPlayer {
  rank: number;
  walletAddress: string;
  wins: number;
  losses: number;
  draws: number;
  score: number;
  xp: number;
  level: number;
}

export interface Leaderboard {
  ranked: RankedPlayer[];
  callerRank: number | null;
}