# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Wallet (SIWE) authentication
  - name: players
    description: Player profile, currency, and inventory
  - name: arena
    description: PvE arena matches
  - name: leaderboard
    description: Ranked player standings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/nonce:
    get:
      operationId: getAuthNonce
      tags: [auth]
      summary: Get a fresh SIWE nonce
      description: Issues a nonce and stores it on the session for the upcoming verify call.
      responses:
        "200":
          description: Nonce issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthNonce"

  /auth/verify:
    post:
      operationId: verifyAuth
      tags: [auth]
      summary: Verify a signed SIWE message and establish a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiweVerification"
      responses:
        "200":
          description: Signature verified, session established
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid nonce or signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/session:
    get:
      operationId: getAuthSession
      tags: [auth]
      summary: Get the current session, if any
      responses:
        "200":
          description: Current session (user is null when not signed in)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"

  /auth/logout:
    post:
      operationId: logoutAuth
      tags: [auth]
      summary: Destroy the current session
      responses:
        "204":
          description: Logged out

  /players/me:
    get:
      operationId: getMyPlayer
      tags: [players]
      summary: Get the signed-in wallet's player row
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /players/me/adjust:
    post:
      operationId: adjustMyPlayer
      tags: [players]
      summary: Apply a gold/xp delta to the signed-in wallet's player row
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerAdjustment"
      responses:
        "200":
          description: Updated player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /players/me/goal:
    patch:
      operationId: updateMyPlayerGoal
      tags: [players]
      summary: Set or clear the signed-in wallet's selected training goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerGoalUpdate"
      responses:
        "200":
          description: Updated player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /players/me/items:
    get:
      operationId: listMyPlayerItems
      tags: [players]
      summary: List the signed-in wallet's inventory
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerItem"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addMyPlayerItem
      tags: [players]
      summary: Add a looted item to the signed-in wallet's inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerItem"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /players/me/items/{instanceId}:
    patch:
      operationId: updateMyPlayerItem
      tags: [players]
      summary: Equip or unequip an owned item
      parameters:
        - name: instanceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerItem"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /arena/match:
    post:
      operationId: recordArenaMatch
      tags: [arena]
      summary: Record a completed arena match for the signed-in wallet
      description: Insert-only — currency rewards must be applied separately via /players/me/adjust.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArenaMatchInput"
      responses:
        "200":
          description: Match recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArenaMatchResult"
        "400":
          description: Missing or invalid fields, or not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get ranked players by arena score
      responses:
        "200":
          description: Leaderboard standings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leaderboard"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    AuthNonce:
      type: object
      required: [nonce]
      properties:
        nonce:
          type: string

    SiweVerification:
      type: object
      required: [message, signature]
      properties:
        message:
          type: string
        signature:
          type: string

    Player:
      type: object
      required: [walletAddress, xp, gold, level]
      properties:
        walletAddress:
          type: string
        xp:
          type: integer
        gold:
          type: integer
        level:
          type: integer
        selectedGoal:
          anyOf:
            - type: string
            - type: "null"

    PlayerGoalUpdate:
      type: object
      required: [selectedGoal]
      properties:
        selectedGoal:
          anyOf:
            - type: string
            - type: "null"

    AuthUser:
      type: object
      required: [walletAddress, player]
      properties:
        walletAddress:
          type: string
        player:
          $ref: "#/components/schemas/Player"

    AuthSession:
      type: object
      required: [user]
      properties:
        user:
          anyOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"

    PlayerAdjustment:
      type: object
      properties:
        goldDelta:
          type: integer
        xpDelta:
          type: integer

    PlayerItem:
      type: object
      required: [instanceId, itemId, slot, equipped, obtainedAt]
      properties:
        instanceId:
          type: string
        itemId:
          type: string
        slot:
          type: string
        equipped:
          type: boolean
        obtainedAt:
          type: string
          format: date-time

    PlayerItemInput:
      type: object
      required: [itemId, slot]
      properties:
        itemId:
          type: string
        slot:
          type: string

    PlayerItemUpdate:
      type: object
      required: [equipped]
      properties:
        equipped:
          type: boolean

    ArenaMatchInput:
      type: object
      required: [opponentId, opponentName, result]
      properties:
        opponentId:
          type: string
        opponentName:
          type: string
        result:
          type: string
          enum: [win, loss, draw]
        xpEarned:
          type: integer
        goldEarned:
          type: integer

    ArenaMatchResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    RankedPlayer:
      type: object
      required: [rank, walletAddress, wins, losses, draws, score, xp, level]
      properties:
        rank:
          type: integer
        walletAddress:
          type: string
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        score:
          type: integer
        xp:
          type: integer
        level:
          type: integer

    Leaderboard:
      type: object
      required: [ranked, callerRank]
      properties:
        ranked:
          type: array
          items:
            $ref: "#/components/schemas/RankedPlayer"
        callerRank:
          anyOf:
            - type: integer
            - type: "null"